cleared warnings in debug mode.

+++ src/interior_point/ipx/src/forrest_tomlin.cc
@@ -254,7 +254,7 @@ void ForrestTomlin::SolvePermuted(Vector& lhs, char trans) {
     assert(U_.cols() == dim_+num_updates);
 
     // Require num_updates elements workspace at end of lhs.
-    assert(lhs.size() >= dim_+num_updates);
+    assert((int)lhs.size() >= dim_+num_updates);
 
     if (trans == 't' || trans == 'T') {
         // Move replaced entries to the end of the pivot sequence and zero out
